+++ Example/PaymentSheet Example/PaymentSheet Example/PaymentSheetTestPlayground.swift
@@ -5,7 +5,7 @@
 //  Created by David Estes on 5/31/23.
 //
 
-import StripePaymentSheet
+@_spi(STP) import StripePaymentSheet
 import SwiftUI
 
 @available(iOS 15.0, *)
@@ -36,6 +36,7 @@ struct PaymentSheetTestPlayground: View {
                 .autocorrectionDisabled()
                 .textInputAutocapitalization(.never)
         }
+        SettingView(setting: enableIos26Binding)
         Group {
             if playgroundController.settings.merchantCountryCode == .US {
                 SettingView(setting: linkEnabledModeBinding)
@@ -302,6 +303,15 @@ struct PaymentSheetTestPlayground: View {
                 playgroundController.settings.uiStyle = .paymentSheet
             }
             playgroundController.settings.integrationType = newIntegrationType
+        }
+    }
+    var enableIos26Binding: Binding<PaymentSheetTestPlaygroundSettings.EnableIOS26Changes> {
+        Binding<PaymentSheetTestPlaygroundSettings.EnableIOS26Changes> {
+            return playgroundController.settings.enableIOS26Changes
+        } set: { newValue in
+            LiquidGlassDetector.allowNewDesign = newValue == .on
+            playgroundController.appearance = PaymentSheet.Appearance()
+            playgroundController.settings.enableIOS26Changes = newValue
         }
     }
 }