+++ StripePaymentSheet/StripePaymentSheet/Source/Internal/API Bindings/Link/STPAPIClient+Link.swift
@@ -614,16 +614,21 @@ extension STPAPIClient {
         cookieStore: LinkCookieStore,
         consumerAccountPublishableKey: String?,
         requestSurface: LinkRequestSurface = .default,
+        consentGranted: Bool? = nil,
         completion: @escaping (Result<ConsumerSession, Error>) -> Void
     ) {
         let endpoint: String = "consumers/sessions/confirm_verification"
 
-        let parameters: [String: Any] = [
+        var parameters: [String: Any] = [
             "credentials": ["consumer_session_client_secret": consumerSessionClientSecret],
             "type": "SMS",
             "code": code,
             "request_surface": requestSurface.rawValue,
         ]
+
+        if let consentGranted {
+            parameters["consent_granted"] = consentGranted
+        }
 
         makeConsumerSessionRequest(
             endpoint: endpoint,
@@ -634,6 +639,26 @@ extension STPAPIClient {
         )
     }
 
+    func updateConsentStatus(
+        consentGranted: Bool,
+        consumerSessionClientSecret: String,
+        consumerPublishableKey: String?,
+        completion: @escaping (Result<EmptyResponse, Error>) -> Void
+    ) {
+        let endpoint: String = "consumers/sessions/consent_update"
+
+        let parameters: [String: Any] = [
+            "credentials": ["consumer_session_client_secret": consumerSessionClientSecret],
+            "consent_granted": consentGranted,
+        ]
+
+        post(
+            resource: endpoint,
+            parameters: parameters,
+            consumerPublishableKey: consumerPublishableKey,
+            completion: completion
+        )
+    }
 }
 
 // TODO(ramont): Remove this after switching to modern bindings.