+++ CHANGELOG.md
@@ -9,6 +9,22 @@
 ### EmbeddedPaymentElement
 * [Added] `opensCardScannerAutomatically` configuration
 
+## 24.22.0 2025-08-27
+
+### StripeCryptoOnramp
+* [Added] Additional APIs and improvements to support the development of the React Native StripeCryptoOnramp SDK.
+
+## 24.21.2 2025-08-27
+
+### PaymentSheet
+* [Changed] Updated translations for Link signup opt-in copy
+
+## 24.21.1 2025-08-26
+
+### PaymentSheet
+* [Changed] Updated copy for Link signup opt-in UI
+>>>>>>> 7d8dc5851a96b9e208d9ad75eb19e7a2ddb273c9
+
 ## 24.21.0 2025-08-25
 
 ### CustomerSheet